removed drone delte
for some reason _this isn't being passed to the lootDrone script (probs something remoteExec related). As the loot markers are shared there's no need to specifically delete this at time of use. It will get cleaned up at the end of the wave.

If this changes in future to be unique it will need to be revisited.