Stuck AI Improvements
Moved Stuck AI checking out of MissionLoop.sqf

Modified it to check for distance to closest player.

Script now checks distance to nearst player then 30 seconds later checks to see if the AI has moved closer to that position. If the AI is not 15 meters closer and is not know to WEST and is not within 20 meter of the player the AI gets deleted.