New PARAMETER: SUPPORT_MENU
Support Menu will not be available from the Bulwark Box until a player locates a Satelite Dish then it will be unlocked for all players and the player whp found it will be rewards with 20 x the Points Multiplier (default 2000 points).

Satellite Dishes keep spawning each wave until one is found as to make it easier to find.

Also modified Params menu formatting.