ARMAKART and Rage Stim fixes
allowDamage was not working sometimes and player / vehicle would still take damage. Switch to using the damage EH / the variable used to make players immune to damage while being revived.

Also, for some reason the infinite ammo wasn't working in my testing so I changed the fucntion to setAmmo.

Added a small nerf to rage stim, when the effect is over your clip is empty.