New location spawn option: List_SpecificPoint
This allows you to pick a specific area to spawn the bulwark by moving the  "Specific Bulwark POS" (specBulwarkLoc) marker on the map in the editor.

Spawn buildings are still subject to the mission Paramters such as BULWARK_LANDRATIO, LOOT_HOUSE_DENSITY, BULWARK_MINSIZE, etc. The mission will hang or error out if these conditions are not met.