Fix: Prevent players quitting to lobby to rejoin wave
Players can quit to the lobby if they die and rejoin the active wave (though they lose any points, etc).

Added a public variable to keep track of which players are in the wave when it begins.

If a player rejoins the wave they will be killed.

Untested if working but doesn't break anything so being committed.

Also. changed endwave to just revive any Incapacitated players.