Fix: Players respawning as soon as they die when they should enter spectator
Occassionally Players are just respawning as soon as they hold "enter" to to die while INCAPACITATED.

I've moved the spawn time check from onPlayerRespawn to onPlayerKilled to make sure we update what the respawn time should be as soon as they die.

I've made sure we're updating the RESPAWN_TIME global variable whenever the respawn time changes.

tested in local MP. respawning and tickets seem to work. Unconfimed if this fixes the bug.