Kill Score adjustment
As players can see the points they get for a kill it would be nice if they noticed that killing different units awards different points. As such I have added point multipliers to the EditMe.

Each hostile level has a different Multiplier and there are multipliers for cars and armour. Currently Cars and Armour award points for killing the vehicle crew, not the vehicle itself.

Going forward it is important to attach a "killPointMulti"variable to any unit that awards points.