Moved Build Phase variable to missionnamespace
For some reason, having this variable attached to the bulwark box was causing problems in the development branch for the Contact DLC map, Livonia.

No reason to have it attached to the Bulwark Box so I moved it to the mission name space.

There is also a commit with a change to the variable in initplayerlocal that previously got submitted with the mag repack feature .