Fix for Medikit not always reviving players.
Medikit was only reviving if the incomming damage was enough to kill the player. It did not take into account if the player was already damaged.

It now checks if the current damage level plus the incomming damage will kill the player.

Also, it wasn't in the player init so I think it didn't work the first time the player went down.